+++ src/relay/mod.rs
@@ -185,9 +185,16 @@ impl JsRelay {
         self.inner.try_connect(**timeout).await.map_err(into_err)
     }
 
-    /// Disconnect from relay and set status to 'Terminated'
+    /// Disconnect from relay and set status to `Terminated`
     pub fn disconnect(&self) {
         self.inner.disconnect()
+    }
+
+    /// Ban relay and set status to `Banned`.
+    ///
+    /// A banned relay can't reconnect again.
+    pub fn ban(&self) {
+        self.inner.ban()
     }
 
     /// Send msg to relay